Move Namespace of ActivityNamespace
